+++ app/helpers/roles_helper.rb
@@ -1,4 +1,4 @@
-# encoding: utf-8
+# frozen_string_literal: true
 
 #  Copyright (c) 2012-2022, Jungwacht Blauring Schweiz. This file is part of
 #  hitobito and licensed under the Affero General Public License version 3
@@ -63,6 +63,15 @@ module RolesHelper
     options = { include_blank: true }
     selected = existing_role(role) || default_role(group)
     options.merge(selected: selected)
+  end
+
+  def terminate_role_link(role)
+    return unless role.terminatable? && can?(:terminate, role)
+
+    link_to(t('roles/terminations.global.title'),
+            new_group_role_termination_path(role_id: role.id, group_id: role.group.id),
+            class: 'btn btn-xs float-right',
+            remote: true)
   end
 
   private