+++ spec/decorators/role_decorator_spec.rb
@@ -13,15 +13,15 @@ describe RoleDecorator, :draper_with_helpers do
 
   let(:decorator) { described_class.new(role) }
 
-  describe '#for_aside' do
+  around do |example|
+    travel_to(today.midnight) { example.run }
+  end
+
+  describe 'outdated roles' do
     let(:tomorrow) { Time.zone.tomorrow }
     let(:title) { node.find_css('i.fa.fa-exclamation-triangle')[0].attr('title') }
 
     subject(:node) { Capybara::Node::Simple.new(decorator.for_aside) }
-
-    around do |example|
-      travel_to(today.midnight) { example.run }
-    end
 
     it 'wraps role#to_s in strong tag wihtout triangle' do
       expect(node).to have_css('strong', text: role.to_s)
@@ -58,6 +58,37 @@ describe RoleDecorator, :draper_with_helpers do
         expect(node).to have_css('i.fa.fa-exclamation-triangle')
         expect(node).to have_css('strong', text: role.to_s)
         expect(title).to eq 'Die Rolle konnte nicht wie geplant per 13.11.2023 aktiviert werden. Falls das Speichern der Rolle diese nicht aktiviert, wende dich bitte an den Support.'
+      end
+    end
+  end
+
+  context 'terminated role' do
+    let(:context) { double('context') }
+    let(:role) { roles(:bottom_member) }
+
+    [:for_aside, :for_history].each do |method|
+      describe "##{method}" do
+        it 'should return the role name' do
+          assert decorator.terminated? == false
+
+          expect(decorator.send(method)).to eq '<strong>Member</strong>'
+        end
+
+        it 'should return the role name and the termination date' do
+          role.write_attribute(:terminated, true)
+          role.deleted_at = today
+
+          expect(decorator.send(method)).to eq "<strong>Member</strong><br><span>Austritt per #{I18n.l(role.terminated_on)}</span>"
+        end
+
+        it 'should return the role name, outdated info and the termination date' do
+          role.write_attribute(:terminated, true)
+          role.delete_on = today
+
+          outdated_and_terminated_text = "<i title=\"Die Rolle konnte nicht wie geplant am 13.11.2023 terminiert werden. Falls das Speichern der Rolle diese nicht terminiert, wende dich bitte an den Support.\" class=\"fa fa-exclamation-triangle\"></i>&nbsp;<strong>Member (Bis 13.11.2023)</strong><br><span>Austritt per 13.11.2023</span>" 
+
+          expect(decorator.send(method)).to eq outdated_and_terminated_text
+        end
       end
     end
   end