+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *  Rollen mit Start-Datum in der Zukunft können erfasst werden (hitobito#2237)
 *  Eigener JSON:API Endpoint für Gruppen (hitobito#2243)
 *  Personendaten können basierend von Spezifikationen im Wagon automatisiert gelöscht werden. Standardmässig abgestellt (#2106)
+*  Rollen können mit dem class attribute `terminatable`, markiert werden, dass sie von der Person selbstständig beendet werden können (hitobito_sac_cas#133)
 
 ## Version 1.30